Moved Menu and LogoutMenuItem to NotesPageTopBarMenu
